+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3957,7 +3957,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M114XDTTX3J1G1ZE9G3AJG28"),
+  id: z.string().default("01M11P4TSB6BW5664QJHP89FVJ"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>