+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-27T08:20:42.765Z",
+    "2026-08-27T13:21:49.218Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {