+++ packages/orq-rc/RELEASES.md
@@ -11959,3 +11959,13 @@ Based on:
 - [typescript v4.15.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.5 - packages/orq-rc
+
+## 2026-08-27 13:24:48
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.0 (2.932.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.6 - packages/orq-rc